requirements.txt
python-dotenv 
langchain
langchain-community
langchain-groq
streamlit
faiss-cpu
pypdf
langchain-openai
tiktoken
pdfminer.six